Move depends definitions into base/api.py as they will likely be used by other missions
